pyproject.toml
[project]
name = "nai-generate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "yada", email = "trojblue@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "dynamicprompts>=0.31.0",
    "gradio>=5.20.0",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]

[project.scripts]
nai-generate = "nai_generate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
